+++ source/graphics/Frustum.h
@@ -32,9 +32,9 @@ public:
 	//Set the number of planes to use for
 	//calculations. This is clipped to
 	//[0,MAX_NUM_FRUSTUM_PLANES]
-	void SetNumPlanes (int num);
+	void SetNumPlanes (size_t num);
 
-	uint GetNumPlanes() const { return m_NumPlanes; }
+	size_t GetNumPlanes() const { return m_NumPlanes; }
 
 	//The following methods return true if the shape is
 	//partially or completely in front of the frustum planes
@@ -43,15 +43,15 @@ public:
 	bool IsSphereVisible (const CVector3D &center, float radius) const;
 	bool IsBoxVisible (const CVector3D &position,const CBound &bounds) const;
 
-	CPlane& operator[](uint idx) { return m_aPlanes[idx]; }
-	const CPlane& operator[](uint idx) const { return m_aPlanes[idx]; }
+	CPlane& operator[](size_t idx) { return m_aPlanes[idx]; }
+	const CPlane& operator[](size_t idx) const { return m_aPlanes[idx]; }
 
 public:
 	//make the planes public for ease of use
 	CPlane m_aPlanes[MAX_NUM_FRUSTUM_PLANES];
 
 private:
-	int m_NumPlanes;
+	size_t m_NumPlanes;
 };
 
 #endif
